Implement progress counters in tcom tools

+++ AprgFileExtractor/AprgFileExtractor/src/AprgFileExtractor.cpp
@@ -9,17 +9,19 @@
 
 using namespace std;
 
+namespace alba
+{
+
 namespace ProgressCounters
 {
 extern int numberOfFilesToBeAnalyzedForExtraction;
 extern int numberOfFilesAnalyzedForExtraction;
 }
 
-namespace alba
-{
 AprgFileExtractor::AprgFileExtractor()
     : m_grepEvaluator("")
 {}
+
 AprgFileExtractor::AprgFileExtractor(string const& condition)
     : m_grepEvaluator(condition)
 {}
@@ -111,16 +113,19 @@ void AprgFileExtractor::extractAllRelevantFilesInThisDirectory(string const& dir
     ProgressCounters::numberOfFilesToBeAnalyzedForExtraction += listOfFiles.size();
     for(string const& filePath: listOfFiles)
     {
-        AlbaWindowsPathHandler extractedPathHandler(filePath);        if(isRecognizedCompressedFile(extractedPathHandler.getExtension()))
+        AlbaWindowsPathHandler extractedPathHandler(filePath);
+        if(isRecognizedCompressedFile(extractedPathHandler.getExtension()))
         {
             extractAllRelevantFilesInThisCompressedFile(extractedPathHandler.getFullPath());
         }
         ProgressCounters::numberOfFilesAnalyzedForExtraction++;
     }
 }
+
 void AprgFileExtractor::extractAllRelevantFilesInThisCompressedFile(string const& filePathOfCompressedFile) const
 {
-    AlbaWindowsPathHandler compressedFilePathHandler(filePathOfCompressedFile);    if(isTheExtensionXz(compressedFilePathHandler.getExtension()))
+    AlbaWindowsPathHandler compressedFilePathHandler(filePathOfCompressedFile);
+    if(isTheExtensionXz(compressedFilePathHandler.getExtension()))
     {
         extractAllFilesRecursively(filePathOfCompressedFile);
     }
@@ -147,10 +152,12 @@ void AprgFileExtractor::extractAllRelevantFilesRecursively(string const& filePat
     ProgressCounters::numberOfFilesToBeAnalyzedForExtraction += filePaths.size();
     for(string const filePath : filePaths)
     {
-        AlbaWindowsPathHandler filePathHandler(filePath);        if(m_grepEvaluator.evaluate(filePathHandler.getFile()))
+        AlbaWindowsPathHandler filePathHandler(filePath);
+        if(m_grepEvaluator.evaluate(filePathHandler.getFile()))
         {
             AlbaWindowsPathHandler extractedPathHandler(extractOneFile(filePathOfCompressedFile, filePath));
-            if(isRecognizedCompressedFile(extractedPathHandler.getExtension()))            {
+            if(isRecognizedCompressedFile(extractedPathHandler.getExtension()))
+            {
                 extractAllRelevantFilesInThisCompressedFile(extractedPathHandler.getFullPath());
                 extractedPathHandler.deleteFile();
             }
@@ -158,8 +165,10 @@ void AprgFileExtractor::extractAllRelevantFilesRecursively(string const& filePat
         ProgressCounters::numberOfFilesAnalyzedForExtraction++;
     }
 }
+
 bool AprgFileExtractor::isTheExtensionXz(string const& extension) const
 {
-    return stringHelper::isEqualNotCaseSensitive("xz", extension);}
+    return stringHelper::isEqualNotCaseSensitive("xz", extension);
+}
 
 }
